Rework OscFunctions 2 class
