+++ src/mixer/basetrackplayer.cpp
@@ -15,6 +15,7 @@
 #include "mixer/playerinfo.h"
 #include "mixer/playermanager.h"
 #include "moc_basetrackplayer.cpp"
+#include "osc/oscfunctions.h"
 #include "track/track.h"
 #include "util/sandbox.h"
 #include "vinylcontrol/defs_vinylcontrol.h"
@@ -34,9 +35,6 @@ inline double trackColorToDouble(mixxx::RgbColor::optional_t color) {
 
 // EveOSC
 extern std::atomic<bool> s_oscEnabled;
-void oscChangedPlayState(
-        const QString& oscGroup,
-        float playstate);
 // EveOSC
 
 BaseTrackPlayer::BaseTrackPlayer(PlayerManager* pParent, const QString& group)
@@ -1036,7 +1034,8 @@ void BaseTrackPlayerImpl::slotPlayToggled(double value) {
     }
     //  EveOSC begin
     if (s_oscEnabled.load()) {
-        oscChangedPlayState(getGroup(), (float)value);
+        OscFunctions oscFunctions(m_pConfig);
+        oscFunctions.oscChangedPlayState(getGroup(), (float)value);
     }
     //  EveOSC end
 }
